fixed events method
