+++ app/models/client.rb
@@ -1,11 +1,12 @@
 class Client < ActiveRecord::Base
   has_many :payload_requests
-  has_many :user_agents, through: :payload_requests
-  has_many :resolutions, through: :payload_requests
-  has_many :requests,    through: :payload_requests
-  has_many :ips,         through: :payload_requests
-  has_many :urls,        through: :payload_requests
-  has_many :referrers,   through: :payload_requests
+  has_many :user_agent_bs, :through => :payload_requests
+  has_many :resolutions,   :through => :payload_requests
+  has_many :requests,      :through => :payload_requests
+  has_many :ips,           :through => :payload_requests
+  has_many :urls,          :through => :payload_requests
+  has_many :referrers,     :through => :payload_requests
+  has_many :events,        :through => :payload_requests
 
   validates :root_url,      presence: true, uniqueness: true
   validates :identifier,    presence: true, uniqueness: true
@@ -13,7 +14,5 @@ class Client < ActiveRecord::Base
   def valid_params(params)
     Client.exists?(params)
   end
-
-
 
 end
