+++ app/models/event.rb
@@ -3,9 +3,11 @@ class Event < ActiveRecord::Base
 
   validates :name,  presence: true
 
-
-  def event_most_recieved_to_least
-    payload_requests.order('event DESC').pluck(:event)
+  def self.event_most_recieved_to_least(payload_requests)
+    events = payload_requests.order('event_id DESC').pluck(:event_id).uniq
+    events.map do |event|
+      Event.find(event).name
+    end
   end
 
 end
