+++ config.ru
@@ -3,4 +3,4 @@ Bundler.require
 
 require File.expand_path('../config/environment',  __FILE__)
 
-run RushHourApp
+run RushHour::Server
