+++ test/test_helper.rb
@@ -10,12 +10,12 @@ require 'capybara/dsl'
 require 'database_cleaner'
 
 DatabaseCleaner.strategy = :truncation, {except: %[public.schema_migrations]}
-Capybara.app = RushHourApp
+Capybara.app = RushHour::Server
 
 module TestHelpers
   include Rack::Test::Methods
   def app     # def app is something that Rack::Test is looking for
-    RushHourApp
+    RushHour::Server
   end
 
   def setup
